+++ lib/Perl/Tidy.pm
@@ -135,7 +135,7 @@ BEGIN {
     # then the Release version must be bumped, and it is probably past time for
     # a release anyway.
 
-    $VERSION = '20240903';
+    $VERSION = '20240903.01';
 } ## end BEGIN
 
 sub DESTROY {